checks for user and verifies on load
